+++ js/navigation.js
@@ -449,7 +449,7 @@ OCA.Analytics.Navigation = {
             success: function (data) {
                 OCA.Analytics.Navigation.buildNavigation(data);
                 OCA.Analytics.reports = data;
-                if (datasetId) {
+                if (datasetId && data.indexOf(data.find(o => o.id === datasetId)) !== -1) {
                     OCA.Analytics.Sidebar.hideSidebar();
                     let navigationItem = document.querySelector('#navigationDatasets [data-id="' + datasetId + '"]');
                     if (navigationItem.parentElement.parentElement.parentElement.classList.contains('collapsible')) {