+++ l10n/de_DE.js
@@ -85,5 +85,7 @@ OC.L10N.register(
             'Column descriptions are used from the dataset and can not be changed on a report level': 'Spaltenbezeichnungen werden vom Datensatz übernommen und können im Bericht nicht geändert werden',
             'Switch to the dataset maintenance from here?': 'Von hier zur Datensatzpflege wechseln?',
             'Dataset maintenance': 'Datensätze pflegen',
+            'Back to reports': 'Zurück zu Berichten',
+            'New dataset': 'Neuer Datensatz'
     },
     '');